fix: 修复统计页UI问题及主题色无效问题
- 统一 MoniCard 圆角为 extraLarge，解决 carousel 卡片与图表卡片圆角不一致
- 增加柱状图 Canvas paddingBottom 至 48f，防止横轴月份标签被底部裁切
- 调整饼图 radius 计算，预留 stroke 边距避免绘制侵入标题区域
- 改用 remember(seedColor.value) 直接调用 dynamicColorScheme，修复主题色切换不生效
- MoniApp 中 Color 构造显式使用 toULong()，确保种子色值正确传递
- ThemeSettings 默认值对齐 DefaultSeedColor.value.toLong()，消除默认值歧义

+++ android/app/src/main/java/com/agguy/moni/app/MoniApp.kt
@@ -57,7 +57,7 @@ fun MoniApp() {
 
     MoniTheme(
         themeMode = themeSettings.themeMode,
-        seedColor = Color(themeSettings.seedColor),
+        seedColor = Color(themeSettings.seedColor.toULong()),
         dynamicColor = themeSettings.dynamicColor
     ) {
         Scaffold(
